# api/openapi.yaml
openapi: 3.1.0
info:
  title: Roots of Brazil — API do Corpus Fundador
  version: 1.1.0
  summary: Camada de acesso somente leitura ao Corpus Fundador v1.1.
  description: 'Implementa a Especificação Conceitual da API v1.1. Serve as 381 entidades e 1.585 relações
    do Corpus Fundador, homologado na Auditoria Sprint 2.


    **Somente leitura.** Não há POST, PUT ou DELETE: o corpus é curado offline (mineração + auditoria)
    e publicado como snapshot versionado, não editado pela API — Especificação Conceitual, Seção 6.


    **Duas origens de dado.** Listagem e detalhe vêm do banco relacional (Ordem 2); navegação entre entidades
    e busca global vêm do banco de grafo (Ordem 3). A divisão está documentada em cada operação.


    **Fidelidade ao corpus.** Toda propriedade de resposta corresponde 1:1 a uma coluna do Dicionário
    de Dados Oficial v1.2 — sem agregação, transformação ou dado derivado além do já calculado no corpus
    (Especificação Conceitual, Seção 7).


    **Confiabilidade sempre visível.** Nenhuma resposta esconde o grau de certeza: `confiabilidade` acompanha
    toda entidade, e `peso`, `fonte`, `pagina` e `observacoes` acompanham toda relação.


    **Acesso.** Aberto, sem chave de API. Limite de 100 requisições por minuto por IP; acima disso, 429.
    CORS liberado para qualquer origem.'
  license:
    name: MIT
    identifier: MIT
  contact:
    name: Roots of Brazil
servers:
- url: http://localhost:8000
  description: Desenvolvimento local.
...
paths:
...
  /v1/relacoes:
    get:
      tags:
      - Relações
      summary: Lista relações
      operationId: listar_relacoes
      description: Acesso direto à camada RELACOES — Especificação Conceitual, Seção 3. Resolvido pelo
        **banco de grafo** (Ordem 3), que guarda peso, confiabilidade e proveniência como propriedades
        nativas da aresta.
      parameters:
      - name: origem_id
        in: query
        required: false
        description: Filtra por ID de origem exato.
        schema:
          type: string
        example: REC-000001
      - name: destino_id
        in: query
        required: false
        description: Filtra por ID de destino exato.
        schema:
          type: string
        example: ING-000010
      - name: tipo_relacao
        in: query
        required: false
        description: Um dos 12 tipos da ontologia. Dicionário v1.2, Seção 21.
        schema:
          type: string
          enum:
          - USA_INGREDIENTE
          - ASSOCIADO_A_POVO
          - CULTIVADO_EM
          - UTILIZA_TECNICA
          - PREPARADO_COM
          - OCORRE_EM
          - ORIGINARIO_DE
          - PATRIMONIO_DE
          - LOCALIZADO_EM_BIOMA
          - DERIVA_DE
          - VARIANTE_REGIONAL
          - SIMILAR_A
        example: USA_INGREDIENTE
      - $ref: '#/components/parameters/Confiabilidade'
      - $ref: '#/components/parameters/Page'
      - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Página de relações.
          headers:
            ETag:
              description: Hash forte do corpo, para revalidação com If-None-Match.
              schema:
                type: string
            Cache-Control:
              description: O corpus muda por versão, não em tempo real.
              schema:
                type: string
            X-RateLimit-Limit:
              description: Requisições permitidas por janela.
              schema:
                type: integer
            X-RateLimit-Remaining:
              description: Requisições restantes na janela.
              schema:
                type: integer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginaRelacao'
        '304':
          description: Não modificado — o ETag enviado ainda vale.
        '400':
          description: Parâmetro de query inválido.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
        '422':
          description: origem_id/destino_id não corresponde a entidade conhecida.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
        '429':
          description: Limite de 100 requisições por minuto por IP excedido.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
          headers:
            Retry-After:
              description: Segundos até a próxima janela.
              schema:
                type: integer
        '500':
          description: Erro não tratado no servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
  /v1/busca:
    get:
      tags:
      - Busca
      summary: Busca global
      operationId: busca_global
      description: Busca textual multi-entidade — Especificação Conceitual, Seção 4. Resolvida pelo **índice
        full-text do banco de grafo** (`objeto_roots_nome_ft`, criado na carga da Ordem 3 sobre `nome_pt`
        e `descricao_pt`). Retorna itens heterogêneos, cada um com `tipo` indicando o catálogo de origem.
      parameters:
      - name: q
        in: query
        required: true
        description: Termo de busca. Obrigatório.
        schema:
          type: string
          minLength: 1
        example: mandioca
      - name: tipos
        in: query
        required: false
        description: Lista separada por vírgula, para restringir os catálogos. Omitido, busca em todos.
        schema:
          type: string
        example: ingrediente,receita
      - $ref: '#/components/parameters/Page'
      - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Resultados heterogêneos, ordenados por relevância.
          headers:
            ETag:
              description: Hash forte do corpo, para revalidação com If-None-Match.
              schema:
                type: string
            Cache-Control:
              description: O corpus muda por versão, não em tempo real.
              schema:
                type: string
            X-RateLimit-Limit:
              description: Requisições permitidas por janela.
              schema:
                type: integer
            X-RateLimit-Remaining:
              description: Requisições restantes na janela.
              schema:
                type: integer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginaBusca'
        '304':
          description: Não modificado — o ETag enviado ainda vale.
        '400':
          description: Parâmetro de query inválido.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
        '429':
          description: Limite de 100 requisições por minuto por IP excedido.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
          headers:
            Retry-After:
              description: Segundos até a próxima janela.
              schema:
                type: integer
        '500':
          description: Erro não tratado no servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
...
components:
  schemas:
...
    Relacao:
      type: object
      description: Aresta da camada RELACOES. Peso, confiabilidade e proveniência são propriedades da
        própria relação — Dicionário v1.2, Seção 20.
      required:
      - origem_id
      - destino_id
      - tipo_relacao
      - observacoes
      - data_criacao
      - peso
      - metodo_calculo_peso
      properties:
        rel_id:
          type:
          - string
          - 'null'
        origem_id:
          type: string
        destino_id:
          type: string
        tipo_relacao:
          type: string
          enum:
          - USA_INGREDIENTE
          - ASSOCIADO_A_POVO
          - CULTIVADO_EM
          - UTILIZA_TECNICA
          - PREPARADO_COM
          - OCORRE_EM
          - ORIGINARIO_DE
          - PATRIMONIO_DE
          - LOCALIZADO_EM_BIOMA
          - DERIVA_DE
          - VARIANTE_REGIONAL
          - SIMILAR_A
        fonte:
          type:
          - string
          - 'null'
        pagina:
          type:
          - string
          - 'null'
        confiabilidade:
          type:
          - string
          - 'null'
          description: 'Grau de certeza da informação. Seção 11. O valor pode trazer texto livre após
            o emoji (ex.: ''🔵 Inferido com cautela''), por isso filtros casam por prefixo.'
        observacoes:
          type: string
        data_criacao:
          type: string
          format: date
        peso:
          type: number
          format: double
          minimum: 0.0
          maximum: 1.0
          description: Peso da relação, de 0.0 a 1.0. Seção 20.
        metodo_calculo_peso:
          type: string
          description: Como o peso foi obtido. Seção 20.2.
      additionalProperties: false
      example:
        rel_id: REL-000001
        origem_id: REC-000001
        destino_id: ING-000010
        tipo_relacao: USA_INGREDIENTE
        fonte: APOSTILA COZINHA/CULINÁRIA BRASILEIRA (Fichas Técnicas); COZINHA BRASILEIRA (Senac-SP,
          2007)
        pagina: ver Aba 7 (citação do ingrediente)
        confiabilidade: 🟢 Confirmado em várias fontes
        observacoes: Reconstruído da junção Receita-Ingrediente do Corpus Fundador v1.0 (Aba 8).
        data_criacao: '2026-08-05'
        peso: 0.95
        metodo_calculo_peso: Inicializado a partir do mapeamento determinístico de confiabilidade (🟢 Confirmado
          em várias fontes), Dicionário v1.2 Seção 20.2
    PaginaRelacao:
      type: object
      description: Página de relações.
      required:
      - total
      - page
      - page_size
      - items
      properties:
        total:
          type: integer
          description: Total de itens que casam com o filtro.
        page:
          type: integer
          minimum: 1
        page_size:
          type: integer
          minimum: 1
          maximum: 100
        items:
          type: array
          items:
            $ref: '#/components/schemas/Relacao'
      additionalProperties: false
      example:
        total: 1
        page: 1
        page_size: 20
        items:
        - rel_id: REL-000001
          origem_id: REC-000001
          destino_id: ING-000010
          tipo_relacao: USA_INGREDIENTE
          fonte: APOSTILA COZINHA/CULINÁRIA BRASILEIRA (Fichas Técnicas); COZINHA BRASILEIRA (Senac-SP,
            2007)
          pagina: ver Aba 7 (citação do ingrediente)
          confiabilidade: 🟢 Confirmado em várias fontes
          observacoes: Reconstruído da junção Receita-Ingrediente do Corpus Fundador v1.0 (Aba 8).
          data_criacao: '2026-08-05'
          peso: 0.95
          metodo_calculo_peso: Inicializado a partir do mapeamento determinístico de confiabilidade (🟢
            Confirmado em várias fontes), Dicionário v1.2 Seção 20.2
    ResultadoBusca:
      type: object
      description: Item heterogêneo da busca global. O campo `tipo` diz de qual catálogo o resultado veio
        — Especificação Conceitual, Seção 4.
      required:
      - id
      - tipo
      - nome
      - confiabilidade
      properties:
        id:
          type: string
          description: Identificador legível, com prefixo da entidade. Dicionário v1.2, Seção 3.
        tipo:
          type: string
          enum:
          - ingrediente
          - receita
          - tecnica
          - povo
          - territorio
          - bioma
          - patrimonio
        nome:
          type: string
          description: Nome em português do objeto.
        descricao:
          type:
          - string
          - 'null'
        slug:
          type: string
        confiabilidade:
          type: string
          description: 'Grau de certeza da informação. Seção 11. O valor pode trazer texto livre após
            o emoji (ex.: ''🔵 Inferido com cautela''), por isso filtros casam por prefixo.'
        score:
          type: number
          format: double
          description: Relevância devolvida pelo índice full-text do grafo.
      additionalProperties: false
      example:
        id: ING-000031
        tipo: ingrediente
        nome: Cebola
        slug: cebola
        confiabilidade: 🟢 Confirmado em várias fontes
        score: 3.14
    PaginaBusca:
      type: object
      description: Página de resultados da busca global.
      required:
      - total
      - page
      - page_size
      - items
      properties:
        total:
          type: integer
          description: Total de itens que casam com o filtro.
        page:
          type: integer
          minimum: 1
        page_size:
          type: integer
          minimum: 1
          maximum: 100
        items:
          type: array
          items:
            $ref: '#/components/schemas/ResultadoBusca'
      additionalProperties: false
      example:
        total: 1
        page: 1
        page_size: 20
        items:
        - id: ING-000031
          tipo: ingrediente
          nome: Cebola
          slug: cebola
          confiabilidade: 🟢 Confirmado em várias fontes
          score: 3.14
    Erro:
      type: object
      description: Formato de erro padrão — Especificação Conceitual, Seção 5.3.
      required:
      - error
      properties:
        error:
          type: object
          required:
          - code
          - message
          - status
          properties:
            code:
              type: string
              enum:
              - INVALID_PARAMETER
              - NOT_FOUND
              - REFERENTIAL_INTEGRITY_ERROR
              - RATE_LIMIT_EXCEEDED
              - INTERNAL_ERROR
            message:
              type: string
            status:
              type: integer
          additionalProperties: false
      additionalProperties: false
      example:
        error:
          code: NOT_FOUND
          message: Objeto com ID 'ING-999999' não encontrado.
          status: 404
  parameters:
    Page:
      name: page
      in: query
      required: false
      description: Página, começando em 1. Seção 5.1.
      schema:
        type: integer
        minimum: 1
        default: 1
      example: 1
    PageSize:
      name: page_size
      in: query
      required: false
      description: Itens por página. Máximo 100. Seção 5.1.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      example: 20
...
    Confiabilidade:
      name: confiabilidade
      in: query
      required: false
      description: 'Grau de certeza da informação. Seção 11. O valor pode trazer texto livre após o emoji
        (ex.: ''🔵 Inferido com cautela''), por isso filtros casam por prefixo.'
      schema:
        type: string
        enum:
        - 🟢 Confirmado em várias fontes
        - 🟡 Confirmado em uma única fonte
        - 🔵 Inferido
        - 🔴 Pendente de validação
      example: 🟢 Confirmado em várias fontes
...